Fix constant debug message about null object
When there isn't a click, the code was returning {} instead of
std::nullopt, which in a variant turns into the default constructor
on the first possibility. So a null shared pointer typed to an object
was getting created, instead of just the nullopt option indiciating
no selection. This fixes that.